Created DBConnection and DBConfig classes for updating Database connection logic, changed DAO and Demonstration classes to work with it.
