+++ src/main/java/jdbctask/DBConfig.java
@@ -1,0 +1,7 @@
+package jdbctask;
+
+public class DBConfig {
+    public static final String DB_URL = "jdbc:postgresql://localhost:5432/my_ticket_service_db";
+    public static final String USERNAME = "postgres";
+    public static final String PASSWORD = "82463755";
+}
